requirements.txt
flask>=3.0.0
line-bot-sdk>=3.10.0
openai>=1.12.0
chromadb>=0.4.22
pandas>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
rank-bm25>=0.2.2
jieba>=0.42.1

